requirements.txt
python-dotenv
langchain
langchain-openai
langgraph
langchain-community # For Chroma, SentenceTransformers
chromadb
sentence-transformers
tiktoken # Often needed for token counting by langchain